london

A list of Simple Features (sf) data.frame objects containing OpenStreetMap polygons, lines, and points for various OpenStreetMap structures in a small part of central London, U.K. (bbox = -0.13, 51.51, -0.11, 51.52). The list includes:

dat_H: 974 non-primary highways as linestrings

dat_HP: 159 primary highways as linestrings

dat_BNR: 1,716 non-residential buildings as polygons

dat_BR: 43 residential buildings as polygons

dat_BC: 67 commerical buildings as polygons

dat_A: 372 amenities as polygons

dat_P: 13 parks as polygons

dat_T: 688 trees as points

dat_RFH: 1 polygon representing Royal Festival Hall

dat_ST: 1 polygon representing 150 Stamford Street

Format

A list of spatial objects

Details

The vignette basic-maps details how these data were downloaded. Note that these internal versions have had all descriptive data removed other than their names, geometries, and their OSM identification numbers.