osmplotr.

Produces customisable images of OpenStreetMap (OSM) data and enables data visualisation using OSM objects. Extracts data using the overpass API. Contains the following functions, data, and vignettes.

Data Functions

extract_osm_objects: Download arbitrary OSM objects

connect_highways: Returns points sequentially connecting list of named highways

Basic Plotting Functions (without data)

add_axes: Overlay longitudinal and latitudinal axes on plot

add_osm_objects: Overlay arbitrary OSM objects

make_osm_map: Automate map production with structures defined in osm_structures

osm_structures: Define structures and graphics schemes for automating map production

osm_basemap: Initiate a ggplot2 object for an OSM map

print_osm_map: Print a map to specified graphics device

Advanced Plotting Functions (with data)

add_osm_groups: Overlay groups of objects using specified colour scheme

add_osm_surface: Overlay data surface by interpolating given data

add_colourbar: Overlay a scaled colourbar for data added with add_osm_surface

Colour Manipulation Functions

adjust_colours: Lighted or darken given colours by specified amount

colour_mat: Generate continuous 2D spatial matrix of colours

Other Functions

get_bbox: return bounding box from input vector

Data

london: OSM Data from a small portion of central London

Vignettes

basic-maps: Describes basics of downloading data and making custom maps

data-maps: Describes how map elements can be coloured according to user-provided data, whether categorical or continuous